fix cmd types
